Inital commit from R to GitHub
Main goal is just to get code up to the repository and update the README with the information from Shlomo's provided README, tho I had ChatGPT try to streamline the information. All of this can be edited later --> will want to add the citation at the top!

Also tried testing code to see if it would run, and unfortunetly no luck with the first two scripts. The first script called data that I did not have, and the 2nd script had an issue with the observation covariates and could not construct UMFs.